En terminado de crear el CRUD de la aplicacion y corregi unos errores menores.

+++ AppPrestamo/AppPrestamo/Model/Pago.cs
@@ -8,6 +8,6 @@ namespace AppPrestamo.Model
         [Indexed]
         public int IdPrestamo { get; set; }
         public int CantidadDePagos { get; set; }
-        public int MontoPagado { get; set; }
+        public float MontoPagado { get; set; }
     }
 }
